Navbar name for page now only editable if page is in navbar
